Fix brillo

+++ src/objects/torus.cpp
@@ -63,18 +63,18 @@ void Torus::loadFaces(int uSteps, int vSteps) {
             face.vertex3 = idx0;
             MaterialProperties properties = getMaterialProperties(MaterialType::Metal);
 
-            face.material.Ka = { (int) (properties.k_a * 0x00), (int) (properties.k_a * 0x58), (int) (properties.k_a * 0xfc) };
-            face.material.Kd = { (int) (properties.k_d * 0x00), (int) (properties.k_d * 0x58), (int) (properties.k_d * 0xfc) };
-            face.material.Ks = { (int) (properties.k_s * 0x00), (int) (properties.k_s * 0x58), (int) (properties.k_s * 0xfc) };
+            face.material.Ka = { properties.k_a * 0x00, properties.k_a * 0x58, properties.k_a * 0xfc };
+            face.material.Kd = { properties.k_d * 0x00, properties.k_d * 0x58, properties.k_d * 0xfc };
+            face.material.Ks = { properties.k_s * 0x00, properties.k_s * 0x58, properties.k_s * 0xfc };
             face.material.Ns = properties.shininess;
             faces.push_back(face);
 
             face.vertex1 = idx3;
             face.vertex2 = idx2; // wrap-around for the quad
             face.vertex3 = idx0;
-            face.material.Ka = { (int) (properties.k_a * 0xfe), (int) (properties.k_a * 0xdd), (int) (properties.k_a * 0x00) };
-            face.material.Kd = { (int) (properties.k_d * 0xfe), (int) (properties.k_d * 0xdd), (int) (properties.k_d * 0x00) };
-            face.material.Ks = { (int) (properties.k_s * 0xfe), (int) (properties.k_s * 0xdd), (int) (properties.k_s * 0x00) };
+            face.material.Ka = { properties.k_a * 0xfe, properties.k_a * 0xdd, properties.k_a * 0x00 };
+            face.material.Kd = { properties.k_d * 0xfe, properties.k_d * 0xdd, properties.k_d * 0x00 };
+            face.material.Ks = { properties.k_s * 0xfe, properties.k_s * 0xdd, properties.k_s * 0x00 };
             face.material.Ns = properties.shininess;
             faces.push_back(face);
         }
